# Disable xdebug for all php processes except code coverage generator