--- a/.php_cs.dist
+++ b/.php_cs.dist
@@ -14,6 +14,7 @@ return \PhpCsFixer\Config::create()
        'phpdoc_summary' => false,
        'phpdoc_align' => false,
        'yoda_style' => false,
+       'is_null' => true,
    ])
    ->setFinder($finder)
 ;