… composer.lock
… tests/Php/PhpIniHelperTest.php
… tests/Php/XdebugHandlerTest.php
… tests/Process/Builder/ProcessBuilderTest.php
… tests/Process/ExecutableFinder/PhpExecutableFinderTest.php