**What I did:**

So I've checked how guys from composer did the same feature and I've done it almost in the same way.

Left condition when we check is xdebug loaded or not. We still need to have this check as we can't determine is xdebug installed or it is just disabled. So I decided to left condition and user still should have xdebug enabled.

Also symfony/process was updated to the latest stable version.

suggestions are welcome.

**What I didn't do:**
We will need to do the same for the native phpdbg. I think it should be separate PR.

p.s. @borNfreee if you don't mind could you please check execution time before and after this enhancement.