@theofidry 2017-11-23
I can't take a look at it right now but it might be worth to hold that for a bit and check what's going turn up in https://github.com/composer/xdebug-handler as there might be something we can use there