@borNfreee 2017-11-23
> if you don't mind could you please check execution time before and after this enhancement.

@sidz Sure, will do it today, thank you!

> and check what's going turn up in https://github.com/composer/xdebug-handler as there might be something we can use there

this looks interesting, but I have several points to discuss:

1. Can we reuse that [standalone package](https://github.com/composer/xdebug-handler/pull/4/files) when it is extracted from composer @sidz ? As I remember, you said that we remove some parts of their code (at least about `version`) and have our special cases that differ from composer's XdebugHandler implementation. 

2. Do we want to have another dependency?