@sidz 2017-11-23
> Can we reuse that standalone package when it is extracted from composer @sidz ? 
Not sure as I don't know what this standalone version will have.

> As I remember, you said that we remove some parts of their code (at least about version) and have our special cases that differ from composer's XdebugHandler implementation.
Yup, you're right we skipped `version` as it not needed in our case. 
Also composer version removes temporary php config right after new command is executed. But for our case we should leave it. (there can be workaround: right now `getCommand` method has `protected` visibility so XdebugHandler class can be easily extended).

> Do we want to have another dependency?
it's hard to say right now. I'd say that we can check what this new standalone package will have and then we can make a decision.

@theofidry we don't know when they will release that library. As for me we can merge all what we have right now as it gives performance boost for users and if it'll be necessary we can rework this part (if it will make sense for us)