@sidz 2017-11-23
@theofidry https://github.com/composer/composer/blob/master/src/Composer/XdebugHandler.php#L36
for example constructor takes OutputInterface and then this class is using in the `getScriptArgs` method which we can't override.

Also we don't need to change these php options: https://github.com/composer/composer/blob/master/src/Composer/XdebugHandler.php#L177

of course we can use Reflection but it's not a good idea.