@borNfreee 2017-11-23
> could you please check execution time before and after this enhancement.

I tried to check it on infection itself using Blackfire, but for some reason blackfire stucks and does nothing for about a minute after all mutant files are generated. I have no idea why it happens. Without Blackfire it works SUPER fast.

So here are results with `time`:

```bash
time infection --threads=4
```

Comparing the `user` time, we have

Run for Project | `master` branch (A), sec | `disable-debugger` (B), sec | Time Decrease A to B | Time Increase B to A
--- | --- | --- | --- | ---
Infection |  148s | 88s | **-40.6%** | **+68%**
JMS Serializer | 465s | 244s | **-48%** | **+90%**
small project | 21s | 11s | **-48%** | **+90%**

So, depending on how we treat the results:

* running infection on itself on this branch takes **40.6%** less time
* running infection on itself on `master` branch takes **68%** more time than on this branch

For JMS Serializer:

* running infection on a JMS Serializer on this branch takes **48%** less time
* running infection on a JMS Serializer on `master` branch takes **90%** more time than on this 

For "small project":

* running infection on a small project on this branch takes **48%** less time
* running infection on a small project on `master` branch takes **90%** more time than on this branch

So it is **1.7x-2x faster** than in `master` branch.  🎉 🎈 🔥 

----------------

This is a huge performance boost. And I also agree with @sidz that we can merge it as is at the moment and when composer package is stable and finalized, we can think about the possibility of using it as a dependency if it will make sense/benefit us.