@sidz 2017-11-26
@theofidry I've applied changes only on files which I touched.