@sidz 2017-11-28
@theofidry could you please re-review it when you'll have a chance?